Add is_unique(), try_unwrap(), try_get_mut() to Rc
Add two new methods to Rc() for manipulating uniquely-owned Rc values.
is_unique() can be used to test if the Rc is uniquely-owned, and
try_unwrap() can remove the value from the Rc if it is uniquely-owned.
try_get_mut() can provide a &mut reference to the contained value if the
Rc is uniquely-owned.

Technically a breaking change, because anyone using an `Rc<T>` where `T`
has methods with these names will be adversely affected if they called
those methods through the implicit dereference of the `Rc`. This is true
for method additions to any smart pointer.

[breaking-change]